requirements.txt
audiomentations==0.27.0
librosa==0.9.2
matplotlib==3.6.2
numpy==1.24.1
pandas==1.5.2
Pillow==9.4.0
python-decouple==3.6
timm==0.6.12
torch==1.13.1
torch_audiomentations==0.11.0
torchaudio==0.13.1
torchmetrics==0.11.0
tqdm==4.64.1
transformers==4.25.1
